requirements.txt
asgiref==3.3.1
Django==3.0.8
django-taggit==1.3.0
mysqlclient @ file:///C:/Users/%EC%B5%9C%EC%98%88%EC%9B%90/Downloads/mysqlclient-1.4.6-cp38-cp38-win32.whl
Pillow==8.1.2
pytz==2021.1
sqlparse==0.4.1
